+++ server/src/feature/embeddedLanguages/modes/getAttributeInterpolationMode.ts
@@ -3,12 +3,11 @@ import { TextDocumentPositionParams } from 'vscode-languageserver';
 
 import { LanguageModelCache } from '../languageModelCache';
 import { LanguageMode, Position, TextDocument } from '../languageModes';
-import { getAureliaVirtualCompletions } from '../../../virtual/virtualCompletion/virtualCompletion';
+import { getAureliaVirtualCompletions } from '../../completions/virtualCompletion';
 import { DefinitionResult } from '../../definition/getDefinition';
 import { aureliaProgram } from '../../../viewModel/AureliaProgram';
-import { getVirtualDefinition } from '../../../virtual/virtualDefinition/virtualDefinition';
 import { getAccessScopeDefinition } from '../../definition/accessScopeDefinition';
-import { VirtualLanguageService } from '../../../virtual/virtualSourceFile';
+import { VirtualLanguageService } from '../../virtual/virtualSourceFile';
 import { getAccessScopeHover } from '../../hover/accessScopeHover';
 
 export function getAttributeInterpolationMode(
